Set ServiceProviderMetadataPath in SAML2 config
sp-metadata.xml is generated by pac4j library. Since v3.4.0 path to the
sp-metadata.xml is required: [1]. gerrit_site/data/<plugin> seems to be
the right directory where this file should be placed.

Given that this plugin is not a real gerrit plugin, but a module,
because it is loaded from gerrit_site/lib directory, we cannot inject
@PluginData path, but have to derive it from SitePaths.data_dir
attribute.

[1] https://github.com/pac4j/pac4j/issues/1210

Change-Id: Ie831d800158df421308153c98ff6489d19be42f5