Adapt assertion consumer service URL setting to pac4j v3
Regarding the callback URL and the SAML support, things have changed
between the versions 2 and 3 of pac4j. In pac4j v2, at the
initialization of the SAML2Client, the callback URL, for example:
http://localhost:8080/callback was turned into
http://localhost:8080/callback?client_name=SAML2Client and the SAML
metadata were generated with that value
(http://localhost:8080/callback?client_name=SAML2Client) as the
entityID.

In pac4j v3, at the initialization of the SAML2Client, the callback URL
is no longer touched and the appropriate callback URL is computed on
demand: either http://localhost:8080/callback?client_name=SAML2Client or
http://localhost:8080/callback/SAML2Client so the SAML metadata use the
original value (http://localhost:8080/callback) as the entityID. This
may cause issue. Nonetheless, we can force this entityID value in the
SAML2 configuration with:

  cfg.setServiceProviderEntityId(
      "http://localhost:8080/callback?client_name=SAML2Client");

See more details in discussion of this PR: [1].

[1] https://github.com/thesamet/gerrit-saml-plugin/pull/18

Change-Id: I399f045b5835c6ef8e8412d61aa8c5bf3adad7e0