Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> > Patch Set 4:
> > 
> > > Patch Set 4:
> > > 
> > > We should create a stable-2.11 branch before this commit
> > 
> > I thought we only care about stable-2.15, not?
> 
> Yes, but because it was already on CI for v2.11 and we still have those plugins available through the Gerrit plugin manager, it would make sense to keep the existing v2.11 build. Other than that, just v2.15 and v2.16 are supported, following Gerrit maintenance policy.

OK, so that we don't need intermediate versions/branches for 2.12, 2.13 and 2.14? But even then, this would make CI side complex, because in 2.11 supported build tool chain is sbt. And starting from 2.15 it is Bazel, and we skipped Buck (and Java 7) for 2.12 and 2.13 versions.

Patch-set: 4
Reviewer: Gerrit User 1011323 <1011323@173816e5-2b9a-37c3-8a2e-48639d4f1153>