Add support for computed display name user property
If identity provider doesn't support multivalue user property in SAML
mapping, allow to compute DisplayName by adding first name and last
name user properties. To achieve that, the following config must be
set:

[saml]
    computedDisplayName = true
    firstNameAttr = firstName
    lastNameAttr = lastName

When identity provider includes in the mapping for the first and last
names:

  firstName=John
  lastName=Doe

then the DisplayName is computed as "John Doe".

Change-Id: I9fe0a4ce1678623dbeae69806fe158d0a29781fd