Update patch set 10
Patch Set 10:

Earlier in this series the in-tree build was done for stable-2.14, and now this change bumps to support 2.16.  Is it safe to assume that the stable-2.14 version also builds and works against Gerrit 2.15?

Should a standalone build be added with explicit API versions?

Patch-set: 10
Reviewer: Gerrit User 1011123 <1011123@173816e5-2b9a-37c3-8a2e-48639d4f1153>