Update patch set 14
Patch Set 14:

What's the criteria for a version being promoted to stable-2.16? We're really looking forward to upgrading our Gerrit instance, but are blocked until there is a 2.16-compatible release of this plugin.

Is the current master release considered safe to use in 2.16, or are there more changes planned before it's ready for general use?

Patch-set: 14