Fix user name misalignment in LDAP and SAML identity provider
With authentication scheme auth.type = HTTP_LDAP the user and group
information is fetched from the LDAP. In case LDAP and SSO SAML
identity provider disagree on user name, the data cannot be fetched
from the LDAP and the user cannot be authenticated.

To rectify the problem allow to optionally fetch the user name from
LDAP with the unique id, that can be requested from the SAML identity
provider as custom SAML attribute in assertion response.

To define the LDAP query in generic way, expose new configuration
option: saml.userNameLdapQuery.

If set, the plugin will try to fetch the user name from the LDAP and
use it for authentication procedure. Note that this will create an
additional load on the LDAP backend and should only be used if the
problem cannot be solved by other means, like align the user names
between different backends or as the very last option, request yet
another SAML attribute in assertion response: "LDAP user name" and
use it instead of using the user name returned by the SAML identity
provider.

Test Plan:

1. Configure this plugin to work with OpenLDAP and Keycloak identity
provider.
2. Configure misaligned user names for a user in OpenLDAP and Keycloak.
3. Configure this option:

saml.userNameLdapQuery = (&(objectClass=person)(entryUUID=${username}))

4. Define the custom field for this user in Keycloak: entryUUID with
the value as seen in LDAP.
5. Verify that authentication works and fetching the user data from
LDAP works, even though the user names are misaligned between Keycloak
and OpenLDAP.

Change-Id: Idd4ee5abcd2190d50c401d92d38305842e6a351f