Update patch set 1
Patch Set 1:

Hi, 

When testing master for 2.16 support, ADFS complains that the authentication request must not contain a NameQualifier, as the SP entity is in the format of nameid-format:entity.

I have tested this change with 2.16 and it is working with ADFS. However, I am not sure what the impact is for Okta/Keycloak. What's the recommended approach for testing? 

Thanks,
Magnus

Patch-set: 1
Reviewer: Gerrit User 1088278 <1088278@173816e5-2b9a-37c3-8a2e-48639d4f1153>