Decode httpDisplaynameHeader to ISO 8859.1
Since [1] Gerrit expects HTTP headers to be encoded as ISO 8859.1[2]
according to RFC 7230[3] the httpDisplaynameHeader is decoded from
ISO 8859.1 since it may contains characters outside of the ASCII
range.
Encode the display name to ISO 8859.1 so that display name isn't
garbled for names that contains characters outside the ASCII
character range.

[1] https://gerrit-review.googlesource.com/c/gerrit/+/94918
[2] https://www.iso.org/standard/28245.html
[3] https://datatracker.ietf.org/doc/html/rfc7230#section-3.2.4

Bug: Issue13076
Change-Id: Id9541d94caec9901049e08de1cd9815a961fc679