Assume that httpDisplaynameHeader is already set correctly
Since [1] Gerrit expects HTTP headers to be encoded as ISO 8859.1[2]
according to RFC 7230[3] the httpDisplaynameHeader is decoded from
ISO 8859.1 since it may contains characters outside of the ASCII
range.
Don't set the httpDisplaynameHeader from AuthenticatedUser, that
is unicode, instead assume that the httpDisplaynameHeader is already
set and that that header is encoded correctly.

[1] https://gerrit-review.googlesource.com/c/gerrit/+/94918
[2] https://www.iso.org/standard/28245.html
[3] https://datatracker.ietf.org/doc/html/rfc7230#section-3.2.4

Bug: Issue13076
Change-Id: Id9541d94caec9901049e08de1cd9815a961fc679