Support display names with non-ISO-8859-1 characters
The SAML plugin relied on request headers to pass the display name to
Gerrit. This was problematic as Gerrit follows RFC-7230's recommendation
[1] of using ISO-8859-1 characters - ideally US-ASCII only. This worked
until change 94918 enforced ISO-8859-1 encoding in the display name
header, and unintentionally (presumably, as there are no tests) also
broke characters that are present but have different encodings in
ISO-8859-1 and UTF-8 (a bug report was raised for this [2]).

This change implements Luca's suggestion [3] not to use headers at all,
but instead make a call via Gerrit's API once the user has been
authenticated. This happens after login and circumvents the issue around
header charsets.

[1] https://datatracker.ietf.org/doc/html/rfc7230#section-3.2.4
[2] https://issues.gerritcodereview.com/issues/296001714
[3] https://gerrit-review.googlesource.com/c/plugins/saml/+/369676/comments/2da3a9f3_1f3c9af3

Bug: Issue 40012596
Change-Id: I321cb9d17d8013a042ea7a4822a606dde8329303