Do not swallow the user provisioning exception
When the user provisioning fails because of the inability
to set its full account name, the error should be also surfaced
to the client, so that it is clear that something isn't right
and its account is not fully processed.

Also cache the filter chain response's status code in case
an error happens when finalising the provisioning of the user.

TODO: Demonstrate with a manual E2E test that failing to update
the full name does result in an error on the client, even with the
full set of components and a real servlet engine.

Change-Id: Ib53f5cfae3f0a2e841e0d575038dd6c182b3bfa9