Do not set full name when not allowed by Realm
Gerrit API for setting the full name may not be allowed
by the authentication Realm, therefore the execution of
the REST-API to set that information from the SAML
plugin may fail.

It is better to avoid setting the full name rather
than having a failed authentication: the underlying Realm
in Gerrit will have a way to populate the missing information.

Change-Id: I4a5687df90c177551cf405ff09e99ab26c659d86