Add tool to create SAML metadata offline
The SAML metadata file (`$SITE/data/saml/sp-metadata.xml`) will be
created on the first login attempt, when the plugin has been installed.
However, at that point authentication would fail until the identity
provider was configured using the metadata file of Gerrit.

To avoid this period in which authentication is not possible, the
metadata can now be created offline.

Change-Id: I0048f7fcf872fc24ed78c36efbf72e182197ccfd