Instantiate SamlClientProvider instead of injecting it
Injecting the SamlClientProvider led to the SamlCLient to be
unconfigured, which broke the plugin.

The assumed reason is that the SamlWebFilter is being injected, when
loading configured filter classes in core, which is happening before
plugin modules are being installed or other plugin classes were
loaded. Thus, injection didn't work properly.

Change-Id: I71796e669cc6064f955a2aca1497561464a21c50