Inject @PluginData Path instead of explicitly building that path
Gerrit core provides the @PluginData Path injection which both
guarantees that this directory will exist before the injection and
honors an already existing plugin data directory, also as a symlink.

Unfortunately, for the plugins provided standalone program
SamlMetadataCreator and for the SamlWebFilterIT test the @PluginData
Path injection is not available. The standalone program doesn't use
injection at all and the AbstractDaemonTest doesn't use the
ServerPluginInfoModule which provides binding for the @PluginData Path.
For that reason this commit copies the logic for the plugin data
directory creatin in the PluginDataDirUtil class and makes use of this
class for these two scenarios. Note that this code is not used for the
lib-module/plugin runtime.

Change-Id: I39cd43a27dec824827bec313b06d0a47271f177c