Add option to enforce auth with IdP on session expiration
If the session in Gerrit expires, it will cause Gerrit to log in with
the SAML IdP again. However, not every IdP will require the user to
authenticate again, e.g. if the auth session in the IdP is longer than
incGerrit. This might especially be the case, if the IdP is used for
multiple applications, where some require less strict authentication.

This change adds an option to enforce authentication on login, i.e. if
the session in Gerrit has expired, users will have to authenticate
again.

Change-Id: Ib558e2b3a896b0e096b7b3ca9593e0d2b5d7b88f